+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
 ## 1.0.5
-
-- ...
+- Added undo pane.
+- Fixed context menu behavior.
+- Added setting to bypass file-update dialogs as per Leo's #4570 and Removed now unused UI file-update settings.
 
 ## 1.0.4